Change routes
Although its not really "pretty", changing the patch/delete routes to get's will make the user experience a lot better. This was, a user fills the form in the website, makes a post request on submit, and an email is sent will the info and the confirmation link. This confirmation link is now a get request, so that the db is updated and the user is redirected to the homepage with a success message. Before this change. there was an extra route in this process, so the user had to fill the form, submit, get an email and fill a new form or click a button. TLDR: One less action from the user.